--- api/monitoring_handler.go
@@ -1,5 +1,0 @@
-package api
-
-import "net/http"
-
-func getHealth(_ http.ResponseWriter, _ *http.Request) {}
