+++ cmd/server/main.go
@@ -2,22 +2,22 @@ package main
 
 import (
 	"context"
+	"github.com/minguu42/mtasks/pkg/app"
+	"github.com/minguu42/mtasks/pkg/logging"
+	"github.com/minguu42/mtasks/pkg/server"
 	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
-
-	"github.com/minguu42/mtasks/api"
-	"github.com/minguu42/mtasks/pkg/logging"
 )
 
 func main() {
-	if err := api.OpenDB(api.DSN("root", "", "mtasks-db-local", 3306, "db_local")); err != nil {
+	if err := app.OpenDB(app.DSN("root", "", "mtasks-db-local", 3306, "db_local")); err != nil {
 		logging.Fatalf("api.OpenDB failed: %v", err)
 	}
-	defer api.CloseDB()
+	defer app.CloseDB()
 
-	s := api.NewServer()
+	s := server.NewServer()
 	shutdownErr := make(chan error, 1)
 	go func() {
 		sigterm := make(chan os.Signal, 1)
