+++ pkg/app/db.go
@@ -1,4 +1,4 @@
-package api
+package app
 
 import (
 	"database/sql"
