+++ pkg/app/error.go
@@ -1,4 +1,4 @@
-package api
+package app
 
 type errorResponse struct {
 	Message string `json:"message"`
