+++ pkg/app/monitoring_handler.go
@@ -1,0 +1,6 @@
+package app
+
+import "net/http"
+
+// GetHealth は GET /health に対応するハンドラ関数
+func GetHealth(_ http.ResponseWriter, _ *http.Request) {}
