+++ pkg/app/task.go
@@ -1,4 +1,4 @@
-package api
+package app
 
 import "time"
